7-evaluation/requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-google-genai>=2.0.0
langsmith
langfuse
openai
python-dotenv
jsonschema
numpy
tqdm
pydantic
